WIP: rename lib and bump versions
 - rename to vision-camera-code-scanner
- bump rn to last version
- bump reanimated
- bump  ndk
- bump vision-camera
- rename project files and dependencies
- fix build issues in android and ios
- fix build issues on apple silicon (M1 users)
