+++ CONTRIBUTING.md
@@ -51,9 +51,9 @@ Remember to add tests for your change if possible. Run the unit tests by:
 yarn test
 ```
 
-To edit the Objective-C files, open `example/ios/VisionCameraQrcodeScannerExample.xcworkspace` in XCode and find the source files at `Pods > Development Pods > vision-camera-qrcode-scanner`.
+To edit the Objective-C files, open `example/ios/VisionCameraCodeScannerExample.xcworkspace` in XCode and find the source files at `Pods > Development Pods > vision-camera-code-scanner`.
 
-To edit the Kotlin files, open `example/android` in Android studio and find the source files at `visioncameraqrcodescanner` under `Android`.
+To edit the Kotlin files, open `example/android` in Android studio and find the source files at `visioncameracodescanner` under `Android`.
 
 ### Commit message convention
