+++ README.md
@@ -1,14 +1,14 @@
-# vision-camera-qrcode-scanner
+# vision-camera-code-scanner
 
 VisionCamera Frame Processor Plugin to read barcodes using MLKit Vision Barcode Scanning
 
 ## Installation
 
 ```sh
-yarn add vision-camera-qrcode-scanner
+yarn add vision-camera-code-scanner
 ```
 
-make sure you correctly [setup](<(https://docs.swmansion.com/react-native-reanimated/docs/fundamentals/installation/)>) react-native-reanimated and insert as a first line of your [`index.tsx`](https://github.com/rodgomesc/vision-camera-qrcode-scanner/blob/1409a8afd02328a26e336036493b2d6ef8441359/example/index.tsx#L1)
+make sure you correctly [setup](<(https://docs.swmansion.com/react-native-reanimated/docs/fundamentals/installation/)>) react-native-reanimated and insert as a first line of your [`index.tsx`](https://github.com/rodgomesc/vision-camera-code-scanner/blob/1409a8afd02328a26e336036493b2d6ef8441359/example/index.tsx#L1)
 
 ```sh
 import 'react-native-reanimated'
@@ -20,7 +20,7 @@ Add this to your `babel.config.js`
 [
   'react-native-reanimated/plugin',
   {
-    globals: ['__scanQRCodes'],
+    globals: ['__scanCodes'],
   },
 ]
 ```
@@ -37,7 +37,7 @@ import * as React from 'react';
 import { StyleSheet, Text } from 'react-native';
 import { useCameraDevices } from 'react-native-vision-camera';
 import { Camera } from 'react-native-vision-camera';
-import { useScanBarcodes, BarcodeFormat } from 'vision-camera-qrcode-scanner';
+import { useScanBarcodes, BarcodeFormat } from 'vision-camera-code-scanner';
 
 export default function App() {
   const [hasPermission, setHasPermission] = React.useState(false);
