+++ example/android/app/src/debug/java/com/example/visioncameracodescanner/ReactNativeFlipper.java
@@ -4,7 +4,7 @@
  * <p>This source code is licensed under the MIT license found in the LICENSE file in the root
  * directory of this source tree.
  */
-package com.example.visioncameraqrcodescanner;
+package com.example.visioncameracodescanner;
 
 import android.content.Context;
 import com.facebook.flipper.android.AndroidFlipperClient;
