+++ example/android/app/src/main/java/com/example/visioncameracodescanner/MainApplication.java
@@ -1,4 +1,4 @@
-package com.example.visioncameraqrcodescanner;
+package com.example.visioncameracodescanner;
 
 import android.app.Application;
 import android.content.Context;
@@ -10,7 +10,7 @@ import com.facebook.react.ReactInstanceManager;
 import com.facebook.soloader.SoLoader;
 import java.lang.reflect.InvocationTargetException;
 import java.util.List;
-import com.visioncameraqrcodescanner.VisionCameraQrcodeScannerPluginPackage;
+import com.visioncameracodescanner.VisionCameraCodeScannerPluginPackage;
 
 import com.facebook.react.bridge.JSIModulePackage;
 import com.swmansion.reanimated.ReanimatedJSIModulePackage;
@@ -28,9 +28,9 @@ public class MainApplication extends Application implements ReactApplication {
         protected List<ReactPackage> getPackages() {
           @SuppressWarnings("UnnecessaryLocalVariable")
           List<ReactPackage> packages = new PackageList(this).getPackages();
-          // Packages that cannot be autolinked yet can be added manually here, for VisionCameraQrcodeScannerExample:
+          // Packages that cannot be autolinked yet can be added manually here, for VisionCameraCodeScannerExample:
           // packages.add(new MyReactNativePackage());
-          packages.add(new VisionCameraQrcodeScannerPluginPackage());
+          packages.add(new VisionCameraCodeScannerPluginPackage());
           return packages;
         }
 
@@ -69,7 +69,7 @@ public class MainApplication extends Application implements ReactApplication {
          We use reflection here to pick up the class that initializes Flipper,
         since Flipper library is not available in release mode
         */
-        Class<?> aClass = Class.forName("com.visioncameraqrcodescannerExample.ReactNativeFlipper");
+        Class<?> aClass = Class.forName("com.visioncameracodescannerExample.ReactNativeFlipper");
         aClass
             .getMethod("initializeFlipper", Context.class, ReactInstanceManager.class)
             .invoke(null, context, reactInstanceManager);
