+++ example/android/gradle.properties
@@ -19,4 +19,4 @@
 
 android.useAndroidX=true
 android.enableJetifier=true
-FLIPPER_VERSION=0.93.0
+FLIPPER_VERSION=0.99.0
