+++ example/android/settings.gradle
@@ -1,6 +1,6 @@
-rootProject.name = 'VisionCameraQrcodeScannerExample'
+rootProject.name = 'VisionCameraCodeScannerExample'
 apply from: file("../node_modules/@react-native-community/cli-platform-android/native_modules.gradle"); applyNativeModulesSettingsGradle(settings)
 include ':app'
 
-include ':visioncameraqrcodescanner'
-project(':visioncameraqrcodescanner').projectDir = new File(rootProject.projectDir, '../../android')
+include ':visioncameracodescanner'
+project(':visioncameracodescanner').projectDir = new File(rootProject.projectDir, '../../android')
