+++ example/app.json
@@ -1,4 +1,4 @@
 {
-  "name": "VisionCameraQrcodeScannerExample",
-  "displayName": "VisionCameraQrcodeScanner Example"
+  "name": "VisionCameraCodeScannerExample",
+  "displayName": "VisionCameraCodeScanner Example"
 }
