+++ example/ios/Podfile
@@ -3,12 +3,12 @@ require_relative '../node_modules/@react-native-community/cli-platform-ios/nativ
 
 platform :ios, '11.0'
 
-target 'VisionCameraQrcodeScannerExample' do
+target 'VisionCameraCodeScannerExample' do
   config = use_native_modules!
 
   use_react_native!(:path => config["reactNativePath"])
 
-  pod 'vision-camera-qrcode-scanner', :path => '../..'
+  pod 'vision-camera-code-scanner', :path => '../..'
 
   # Enables Flipper.
   #
@@ -17,5 +17,11 @@ target 'VisionCameraQrcodeScannerExample' do
   use_flipper!()
   post_install do |installer|
     react_native_post_install(installer)
+    __apply_Xcode_12_5_M1_post_install_workaround(installer)
+    
+    installer.pods_project.build_configurations.each do |config|
+      config.build_settings['ARCHS[sdk=iphonesimulator*]'] =  `uname -m`
+    end
+    
   end
 end
