+++ example/ios/VisionCameraCodeScannerExample-Bridging-Header.h
