+++ example/ios/VisionCameraCodeScannerExample.xcworkspace/contents.xcworkspacedata
@@ -2,7 +2,7 @@
 <Workspace
    version = "1.0">
    <FileRef
-      location = "group:VisionCameraQrcodeScannerExample.xcodeproj">
+      location = "group:VisionCameraCodeScannerExample.xcodeproj">
    </FileRef>
    <FileRef
       location = "group:Pods/Pods.xcodeproj">
