+++ example/ios/VisionCameraCodeScannerExample.xcworkspace/xcshareddata/IDEWorkspaceChecks.plist
