+++ example/ios/VisionCameraCodeScannerExample/AppDelegate.h
