+++ example/ios/VisionCameraCodeScannerExample/AppDelegate.m
@@ -38,7 +38,7 @@ static void InitializeFlipper(UIApplication *application) {
   #endif
   RCTBridge *bridge = [[RCTBridge alloc] initWithDelegate:self launchOptions:launchOptions];
   RCTRootView *rootView = [[RCTRootView alloc] initWithBridge:bridge
-                                                   moduleName:@"VisionCameraQrcodeScannerExample"
+                                                   moduleName:@"VisionCameraCodeScannerExample"
                                             initialProperties:nil];
 
   rootView.backgroundColor = [[UIColor alloc] initWithRed:1.0f green:1.0f blue:1.0f alpha:1];
