+++ example/ios/VisionCameraCodeScannerExample/Images.xcassets/AppIcon.appiconset/Contents.json
