+++ example/ios/VisionCameraCodeScannerExample/Images.xcassets/Contents.json
