+++ ios/VisionCameraCodeScanner.m
@@ -1,0 +1,5 @@
+#import <Foundation/Foundation.h>
+#import <VisionCamera/FrameProcessorPlugin.h>
+
+@interface VISION_EXPORT_SWIFT_FRAME_PROCESSOR(scanCodes, VisionCameraCodeScanner)
+@end
