--- ios/VisionCameraQrcodeScanner.m
@@ -1,5 +1,0 @@
-#import <Foundation/Foundation.h>
-#import <VisionCamera/FrameProcessorPlugin.h>
-
-@interface VISION_EXPORT_SWIFT_FRAME_PROCESSOR(scanQRCodes, VisionCameraQrcodeScanner)
-@end
